+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/blocks/MovingSlidyBlockSource.java
@@ -20,6 +20,7 @@ public class MovingSlidyBlockSource extends DirectionalBlock {
         super(properties);
     }
 
+
     @Override
     protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
         super.createBlockStateDefinition(builder);
@@ -41,10 +42,12 @@ public class MovingSlidyBlockSource extends DirectionalBlock {
         BlockState frontState = level.getBlockState(pos.relative(state.getValue(FACING)));
         if (!frontState.is(ModRegistry.MOVING_SLIDY_BLOCK.get()) ||
                 frontState.getValue(FACING) != state.getValue(FACING)) {
-            level.setBlock(pos, Blocks.AIR.defaultBlockState(), Block.UPDATE_KNOWN_SHAPE | Block.UPDATE_ALL | Block.UPDATE_MOVE_BY_PISTON);
+            level.setBlock(pos, Blocks.AIR.defaultBlockState(),
+                    Block.UPDATE_ALL | Block.UPDATE_MOVE_BY_PISTON);
+            //mixin suppresses obserber updates
+            //we need to update all for falling blocks
         }
     }
-    //TODO: fix not updating neighboring blocks
 
 
 }